:recycle: Small fixes
