+++ src/main/kotlin/com/github/sloppylopez/moneypennyideaplugin/toolWindow/ToolWindowFactory.kt
@@ -1,10 +1,17 @@
 package com.github.sloppylopez.moneypennyideaplugin.toolWindow
 
+import com.github.sloppylopez.moneypennyideaplugin.actions.SendToPromptFileFolderTreeAction
+import com.github.sloppylopez.moneypennyideaplugin.actions.SendToPromptTextEditorAction
 import com.intellij.openapi.diagnostic.Logger
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.wm.ToolWindow
 import com.intellij.openapi.wm.ToolWindowFactory
 import com.github.sloppylopez.moneypennyideaplugin.helper.ToolWindowHelper.Companion.addTabbedPaneToToolWindow
+import com.github.sloppylopez.moneypennyideaplugin.intentions.RefactorIntentionFactory
+import com.intellij.openapi.application.ApplicationManager
+import com.intellij.openapi.application.ModalityState
+import com.intellij.openapi.components.service
+import javax.swing.SwingUtilities
 
 class ToolWindowFactory : ToolWindowFactory {
     override fun createToolWindowContent(
@@ -12,7 +19,21 @@ class ToolWindowFactory : ToolWindowFactory {
         toolWindow: ToolWindow
     ) {
         try {
-            addTabbedPaneToToolWindow(project, toolWindow)
+            val sendToPromptFileFolderTreeAction = SendToPromptFileFolderTreeAction(project)
+            sendToPromptFileFolderTreeAction.registerFolderTreeAction()
+            val sendToPromptTextEditorAction = SendToPromptTextEditorAction(project)
+            sendToPromptTextEditorAction.registerFileEditorAction()
+            val refactorIntentionFactory = project.service<RefactorIntentionFactory>()
+            SwingUtilities.invokeLater {
+                ApplicationManager.getApplication().invokeLater(
+                    {
+                        refactorIntentionFactory.removeIntentions()
+                        refactorIntentionFactory.addIntentionToAllEditors()
+                    },
+                    ModalityState.NON_MODAL
+                )
+            }
+            addTabbedPaneToToolWindow(project, toolWindow,)
         } catch (e: Exception) {
             Logger.getInstance("ToolWindowFactory").error(e.stackTraceToString())
         }
